backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.4
whitenoise==6.11.0
djangorestframework
djangorestframework-simplejwt
Pillow
torch
torchvision

